+++ src/shared/util/typeGenerator/json.ts
